Make menu button also optional as extra ConfigList feature
The 'OK' button and the 'MENU' button have the same functionality as it
both gives the choicebox on selection lists. I suggest the 'OK' button
is the prefered method as it also does toggle switches etc. So I gave
the 'MENU' button back as extra feature so you can for example put a
config behind a config in an easy way. The syntax is the same as I did
with the blue and yellow buttons.

Co-Authored-By: littlesat <littlesat99@yahoo.com>