In setup when the setup list seems to be empty then indicate this
Previously when a setup list seems to be empty e.g. due to conditions
then we get an empty list. This is a rare situation but at least it is
not arranged that we at least see one ConfigNothing tiems with a
description identifying that there are no config items available instead
if showing an empty list without any identification why it does happen

Co-Authored-By: littlesat <littlesat99@yahoo.com>