Resolve exit key functionality in SoftcamSetup
As the default from the config was not set configList cannot determine
if a value was changed. This resolved it

I also removed the embedded screen (sounds I prevuously forgot to commit
this part) so sorry that this is also include. The screen cannot always
fallback to the standard SetupScreen when it is not available.

Co-Authored-By: littlesat <littlesat99@yahoo.com>