[SkinSelection.py] Add a new module for skin selection and management
SkinSelection is a new skin selection module/screen that is based on the Setup screen.  This module/screen replaces the SkinSelector module/screen.  At the moment this new screen manages both the GUI and LCD skin selection.  The code is designed to support future skin management features currently being developed.

SkinSelection uses a skin identical to the Setup skin with the addition of an extra widget "preview" to support the display of the skin preview images.

Co-Authored-By: IanSav <IanSav@users.noreply.github.com>