[ImportChannels] When network is offline do not do recursion in Impor… …tchannels
At the function getUrl in Importchannels when the network was offline
recursion was performed. This means when the network for some reason
does never become online (e.g. box at a camping) it will be recusrsive
forever which means enigma2 will crash at a specific time. Therefore
retrying to check in a loop might be a better approach

cherry-pick > OpenPLi/enigma2@1776d66