[ImportChannels] Change the order of import information in ImportChan… …nels
First the terristical type was read, then the epg and as last the
channels. This change does change the order by first reading the
channels, then the epg and at last the terristical type. In case there
is an error I suggest this is a better order.

cherry-pick > OpenPLi/enigma2@0c33bb8