In ServiceList remove a redudant line of code that caused glitches
This line of code is redundant as this is also done in setMode. Doing
this shortly after eachother causes glitches in the ChannelSelection
list on at least the VU+ Ultimo 4K where a channel name or event name
became slightly dislocated.

It took me a long time to find this 'glitch'

Co-Authored-By: littlesat <littlesat99@yahoo.com>