Optionally add MovieSelection to Main Menu (#3539)
* Optionally add MovieSelection to Main Menu

This change adds an option "Show Movie Selection Player" to the "OSD Settings" screen that will allow users to add a ""Movie Selection Player" option to the Main Menu.  This is useful for users who do not have the appropriate FILE, LIST, MEDIA, PVR, or VIDEO button on their remote control.  The default is to not add this menu item to the Main Menu.

---------

Co-Authored-By: Ian <IanSav@users.noreply.github.com>