Try to bring back seek for Video On Demand in Bouquets in a better way
Standard the numbers are associated to numberzap but with long-exit
button in a seekable service you can from now on toggle between
numberzap and seek stepskip when a service is e.g. seekable. After a zap
it defaults back always to number zap. On the screen you also see
shortly the status if Numberbutton is zap or in seekmode. For 'normal'
services the long exit button does nothing.

Co-Authored-By: littlesat <littlesat99@yahoo.com>